added a check to see if the script was already installed, cleaned up oh my zsh install, added installer for OSX-ElCap cursors

+++ install.sh
@@ -6,11 +6,34 @@ if [ "$(id -u)" -eq 0 ]; then
 	exit
 fi
 
+# Already installed check
+if [ "${1}" != '--force-install' ] && [ -f ".installed" ]; then 
+	echo "Config might've already been installed. Run again with \"--force-install\" to force the script to run."
+	exit
+fi
+touch .installed
+
 # Dependencies
-sudo apt install -y coreutils micro python3 wget zoxide zsh
+sudo apt install -y curl coreutils micro python3 pipx wget zoxide zsh
 
 # Oh My Zsh
 cp ./dotfiles/.zshrc ~/.zshrc
-wget --directory-prefix=downloads --output-document=zsh-install.sh \
-	"https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh"
-KEEP_ZSHRC=yes sh downloads/zsh-install.sh
+wget -O- "https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh" | \
+	KEEP_ZSHRC=yes sh
+
+# OSX El Capitan Cursors by XXMMTRTXX
+# https://www.mate-look.org/p/1084939
+sudo mkdir -p /usr/share/icons/OSX-ElCapitan
+
+# https://unix.stackexchange.com/a/743991
+# this took me forever
+# get JSON file that contains download info
+curl -Lfs "https://www.mate-look.org/p/1084939/loadFiles" | \
+	# parse that JSON, get the most recent version's download url (in this case, the last returned item)
+	jq -r '.files | .[-1] | .url' | \
+	# parse junk characters
+	perl -pe 's/\%(\w\w)/chr hex $1/ge' | \
+	# download the file
+	xargs wget -O- | \
+	# extract the cursors to the global icons folder
+	sudo tar -xjf - -C /usr/share/icons/OSX-ElCapitan --strip-components=2
